+++ src/body/Stash Tag Image Cropper.user.js
@@ -36,7 +36,9 @@
                 cropBtnContainer.setAttribute("id", cropBtnContainerId);
                 cropBtnContainer.classList.add('mb-2', 'text-center');
                 insertAfter(cropBtnContainer, image.parentElement);
-    
+
+                const cropInfo = document.createElement('p');
+
                 const imageUrl = getElementByXpath("//div[contains(@class, 'logo-container')]//img[@class='logo']/@src").nodeValue;
                 const cropStart = document.createElement('button');
                 cropStart.setAttribute("id", "crop-start");
@@ -58,6 +60,9 @@
                         zoomOnWheel: false,
                         ready() {
                             cropAccept.style.display = 'inline-block';
+                        },
+                        crop(e) {
+                            cropInfo.innerText = `X: ${Math.round(e.detail.x)}, Y: ${Math.round(e.detail.y)}, Width: ${Math.round(e.detail.width)}px, Height: ${Math.round(e.detail.height)}px`;
                         }
                     });
                 });
@@ -72,6 +77,7 @@
                     cropStart.style.display = 'inline-block';
                     cropAccept.style.display = 'none';
                     cropCancel.style.display = 'none';
+                    cropInfo.innerText = '';
     
                     const reqData = {
                         "operationName": "TagUpdate",
@@ -102,12 +108,15 @@
                     cropStart.style.display = 'inline-block';
                     cropAccept.style.display = 'none';
                     cropCancel.style.display = 'none';
+                    cropInfo.innerText = '';
     
                     cropper.destroy();
                 });
                 cropBtnContainer.appendChild(cropCancel);
                 cropAccept.style.display = 'none';
                 cropCancel.style.display = 'none';
+
+                cropBtnContainer.appendChild(cropInfo);
             }
         });
     });